move to_data and to_funsor
